backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.7
python-jose==3.3.0
PyJWT==2.8.0
google-genai
pytest==7.4.2
Faker==19.6.2
pandas==2.1.1
python-docx==0.8.11
openpyxl==3.1.2
